Time: 326 ms (100.00%), Space: 75.8 MB (44.44%) - LeetHub

+++ 2530-maximal-score-after-applying-k-operations/2530-maximal-score-after-applying-k-operations.cpp
@@ -14,7 +14,7 @@ public:
         }
         while(!pq.empty() and k>0)
         {
-            int top=pq.top();
+            long long top=pq.top();
             pq.pop();
             
             ans+=top;
